+++ Tests/MapboxMapsTests/Viewport/Mocks/MockViewportStatusObserver.swift
@@ -1,4 +1,4 @@
-@_spi(Experimental) import MapboxMaps
+import MapboxMaps
 
 final class MockViewportStatusObserver: ViewportStatusObserver {
     struct ViewportStatusDidChangeParams: Equatable {